Update Canvas, Annotator
