+++ frontend/src/AuthenticatedApp/paperSetup/components/modals/ViewScriptModal.tsx
@@ -6,7 +6,7 @@ import { ScriptData, ScriptListData } from "backend/src/types/scripts";
 import { Button, Dialog, DialogContent, DialogTitle } from "@material-ui/core";
 import TogglePageComponent from "../../../../components/misc/TogglePageComponent";
 import Canvas from "../../../../components/Canvas";
-import { CanvasMode } from "../../../../types/canvas";
+import { CanvasMode } from "../../../../components/Canvas/types";
 import LoadingSpinner from "../../../../components/LoadingSpinner";
 import DialogTitleWithCloseButton from "../../../../components/dialogs/DialogTitleWithCloseButton";
