--- frontend/src/types/canvas.ts
@@ -1,5 +1,0 @@
-export enum CanvasMode {
-  Pen = "pen",
-  Eraser = "eraser",
-  View = "view"
-}
